Add files via upload
The first and final version, since i don't intend on working further on this, it was a fun little project that allowed me to learn how to use keyboard events in Java and how to work with JFrames and render moving objects. Its incomplete as it lacks the end screen and a scoring system but the game play is smooth and works.